chore(worker): add storage_sync module to storage crate
