+++ worker/crates/storage/src/storage.rs
@@ -1,4 +1,3 @@
-use std::io::Read;
 use std::ops::Range;
 use std::path::Path;
 use std::pin::Pin;
@@ -13,7 +12,6 @@ use object_store::GetResult;
 use object_store::GetResultPayload;
 use object_store::ObjectMeta;
 use object_store::Result;
-use opendal::BlockingReader;
 use opendal::Metakey;
 use opendal::Operator;
 use opendal::Reader;
@@ -22,8 +20,8 @@ use reearth_flow_common::uri::Uri;
 
 #[derive(Debug)]
 pub struct Storage {
-    base_uri: Uri,
-    inner: Operator,
+    pub(crate) base_uri: Uri,
+    pub(crate) inner: Operator,
 }
 
 impl std::fmt::Display for Storage {
@@ -40,18 +38,6 @@ impl Storage {
         }
     }
 
-    pub fn put_sync(&self, location: &Path, bytes: Bytes) -> Result<()> {
-        let p = location.to_str().ok_or(object_store::Error::InvalidPath {
-            source: object_store::path::Error::InvalidPath {
-                path: format!("{:?}", location).into(),
-            },
-        })?;
-        self.inner
-            .blocking()
-            .write(p, bytes)
-            .map_err(|err| format_object_store_error(err, p))
-    }
-
     pub async fn put(&self, location: &Path, bytes: Bytes) -> Result<()> {
         let p = location.to_str().ok_or(object_store::Error::InvalidPath {
             source: object_store::path::Error::InvalidPath {
@@ -62,23 +48,6 @@ impl Storage {
             .write(p, bytes)
             .await
             .map_err(|err| format_object_store_error(err, p))
-    }
-
-    pub fn create_dir_sync(&self, location: &Path) -> Result<()> {
-        let p = location.to_str().ok_or(object_store::Error::InvalidPath {
-            source: object_store::path::Error::InvalidPath {
-                path: format!("{:?}", location).into(),
-            },
-        })?;
-        let p = if !p.ends_with('/') {
-            format!("{}/", p)
-        } else {
-            p.to_string()
-        };
-        self.inner
-            .blocking()
-            .create_dir(p.as_str())
-            .map_err(|err| format_object_store_error(err, p.as_str()))
     }
 
     pub async fn create_dir(&self, location: &Path) -> Result<()> {
@@ -98,23 +67,6 @@ impl Storage {
             .map_err(|err| format_object_store_error(err, p.as_str()))
     }
 
-    pub fn append_sync(&self, location: &Path, bytes: Bytes) -> Result<()> {
-        let p = location.to_str().ok_or(object_store::Error::InvalidPath {
-            source: object_store::path::Error::InvalidPath {
-                path: format!("{:?}", location).into(),
-            },
-        })?;
-        let mut w = self
-            .inner
-            .blocking()
-            .writer_with(p)
-            .append(true)
-            .call()
-            .map_err(|err| format_object_store_error(err, p))?;
-        w.write(bytes)
-            .map_err(|err| format_object_store_error(err, p))
-    }
-
     pub async fn append(&self, location: &Path, bytes: Bytes) -> Result<()> {
         let p = location.to_str().ok_or(object_store::Error::InvalidPath {
             source: object_store::path::Error::InvalidPath {
@@ -130,27 +82,6 @@ impl Storage {
         w.write(bytes)
             .await
             .map_err(|err| format_object_store_error(err, p))
-    }
-
-    pub fn get_sync(&self, location: &Path) -> Result<Bytes> {
-        let p = location.to_str().ok_or(object_store::Error::InvalidPath {
-            source: object_store::path::Error::InvalidPath {
-                path: format!("{:?}", location).into(),
-            },
-        })?;
-        let mut r = self
-            .inner
-            .blocking()
-            .reader(p)
-            .map_err(|err| format_object_store_error(err, p))?;
-
-        let mut buf = Vec::new();
-        r.read_to_end(&mut buf)
-            .map_err(|err| object_store::Error::Generic {
-                store: "IoError",
-                source: Box::new(err),
-            })?;
-        Ok(Bytes::from(buf))
     }
 
     pub async fn get(&self, location: &Path) -> Result<GetResult> {
@@ -183,18 +114,6 @@ impl Storage {
             range: (0..meta.size),
             meta,
         })
-    }
-
-    pub fn exists_sync(&self, location: &Path) -> Result<bool> {
-        let p = location.to_str().ok_or(object_store::Error::InvalidPath {
-            source: object_store::path::Error::InvalidPath {
-                path: format!("{:?}", location).into(),
-            },
-        })?;
-        self.inner
-            .blocking()
-            .is_exist(p)
-            .map_err(|err| format_object_store_error(err, p))
     }
 
     pub async fn exists(&self, location: &Path) -> Result<bool> {
@@ -246,19 +165,6 @@ impl Storage {
         })
     }
 
-    pub fn delete_sync(&self, location: &Path) -> Result<()> {
-        let p = location.to_str().ok_or(object_store::Error::InvalidPath {
-            source: object_store::path::Error::InvalidPath {
-                path: format!("{:?}", location).into(),
-            },
-        })?;
-        self.inner
-            .blocking()
-            .delete(p)
-            .map_err(|err| format_object_store_error(err, p))?;
-        Ok(())
-    }
-
     pub async fn delete(&self, location: &Path) -> Result<()> {
         let p = location.to_str().ok_or(object_store::Error::InvalidPath {
             source: object_store::path::Error::InvalidPath {
@@ -271,41 +177,6 @@ impl Storage {
             .map_err(|err| format_object_store_error(err, p))?;
 
         Ok(())
-    }
-
-    pub fn list_sync(&self, prefix: Option<&Path>, recursive: bool) -> Result<Vec<Uri>> {
-        let p = prefix.ok_or(object_store::Error::InvalidPath {
-            source: object_store::path::Error::InvalidPath {
-                path: format!("{:?}", prefix).into(),
-            },
-        })?;
-        let path =
-            p.to_str()
-                .map(|v| format!("{}/", v))
-                .ok_or(object_store::Error::InvalidPath {
-                    source: object_store::path::Error::InvalidPath {
-                        path: format!("{:?}", prefix).into(),
-                    },
-                })?;
-        let ds = self
-            .inner
-            .blocking()
-            .lister_with(&path)
-            .recursive(recursive)
-            .metakey(Metakey::ContentLength | Metakey::LastModified)
-            .call()
-            .map_err(|err| format_object_store_error(err, ""))?;
-        let result = ds
-            .filter_map(|entry| match entry {
-                Ok(v) => Some(Uri::for_test(&format!(
-                    "{}/{}",
-                    self.base_uri.protocol().as_str_with_separator(),
-                    v.path()
-                ))),
-                Err(_) => None,
-            })
-            .collect::<Vec<_>>();
-        Ok(result)
     }
 
     pub async fn list(
@@ -359,24 +230,6 @@ impl Storage {
             .collect::<Vec<_>>())
     }
 
-    pub fn copy_sync(&self, from: &Path, to: &Path) -> Result<()> {
-        let from = from.to_str().ok_or(object_store::Error::InvalidPath {
-            source: object_store::path::Error::InvalidPath {
-                path: format!("{:?}", from).into(),
-            },
-        })?;
-        let to = to.to_str().ok_or(object_store::Error::InvalidPath {
-            source: object_store::path::Error::InvalidPath {
-                path: format!("{:?}", to).into(),
-            },
-        })?;
-        self.inner
-            .blocking()
-            .copy(from.as_ref(), to.as_ref())
-            .map_err(|err| format_object_store_error(err, from))?;
-        Ok(())
-    }
-
     pub async fn copy(&self, from: &Path, to: &Path) -> Result<()> {
         let from = from.to_str().ok_or(object_store::Error::InvalidPath {
             source: object_store::path::Error::InvalidPath {
@@ -396,7 +249,7 @@ impl Storage {
     }
 }
 
-fn format_object_store_error(err: opendal::Error, path: &str) -> object_store::Error {
+pub(crate) fn format_object_store_error(err: opendal::Error, path: &str) -> object_store::Error {
     use opendal::ErrorKind;
     match err.kind() {
         ErrorKind::NotFound => object_store::Error::NotFound {
@@ -433,26 +286,6 @@ impl Stream for OpendalReader {
                 store: "IoError",
                 source: Box::new(err),
             })
-    }
-}
-
-struct OpendalBlockingReader {
-    inner: BlockingReader,
-}
-
-impl Stream for OpendalBlockingReader {
-    type Item = Result<Bytes>;
-
-    fn poll_next(mut self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        let result = match self.inner.next() {
-            Some(Ok(v)) => Some(Ok(v)),
-            Some(Err(e)) => Some(Err(object_store::Error::Generic {
-                store: "IoError",
-                source: Box::new(e),
-            })),
-            None => None,
-        };
-        Poll::Ready(result)
     }
 }
